% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_dummies}
\alias{example_dummies}
\title{A set of example dummy alignment genes.}
\format{
A data frame with eight rows and three variables:
\describe{
\item{molecule}{the genome}
\item{start}{the start position of the dummy}
\item{end}{the end position of the dummy}
\item{gene}{the name of the dummy gene}
}
}
\usage{
example_dummies
}
\description{
Dummy genes, similar to those generated with \code{make_alignment_dummies()}, for
example purposes only.
}
\keyword{datasets}
