% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_features}
\alias{example_features}
\alias{example_features_polar}
\title{A set of example genetic features.}
\format{
A data frame with 23 rows and five variables:
\describe{
\item{molecule}{the genome}
\item{name}{the name of the feature}
\item{type}{the type of the feature}
\item{position}{the position of the feature}
\item{forward}{is the feature oriented, and if so in the forward direction?}
}

An object of class \code{data.frame} with 6 rows and 5 columns.
}
\usage{
example_features

example_features_polar
}
\description{
Genetic features for example purposes only.
}
\keyword{datasets}
