% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_genes}
\alias{example_genes}
\title{A set of example genes.}
\format{A data frame with 118 rows and four variables:
\describe{
  \item{molecule}{the genome}
  \item{start}{the start position of the gene}
  \item{end}{the end position of the gene}
  \item{gene}{the name of the gene}
  \item{strand}{the strand of the gene}
}}
\usage{
example_genes
}
\description{
Genes for example purposes only.
}
\keyword{datasets}
