% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_terminators}
\alias{example_terminators}
\title{A set of example terminators.}
\format{
A data frame with two rows and three variables:
\describe{
\item{molecule}{the genome}
\item{name}{the name of the terminator}
\item{position}{the position of the terminator}
}
}
\usage{
example_terminators
}
\description{
Terminators for example purposes only.
}
\keyword{datasets}
