% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{feature_garden}
\alias{feature_garden}
\title{The complete set of implemented SBOL sequence features}
\format{
A data frame with four rows and five variables:
\describe{
\item{feature}{the type of sequence feature}
\item{variant}{the variant of the sequence feature}
\item{molecule}{the genome}
\item{start}{the start position or location of the feature}
\item{end}{the end position of the feature}
\item{forward}{whether the feature is on the forward strand}
}
}
\usage{
feature_garden
}
\description{
Sequence features for example purposes only.
}
\keyword{datasets}
