% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_aptamer.R
\name{geom_aptamer}
\alias{geom_aptamer}
\title{A 'ggplot2' geom to draw aptamers}
\usage{
geom_aptamer(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = FALSE,
  height = grid::unit(3, "mm"),
  ...
)
}
\arguments{
\item{mapping, data, stat, position, na.rm, show.legend, inherit.aes, ...}{As
standard for ggplot2. inherit.aes is set to FALSE by default.}

\item{height}{\code{grid::unit()} opbject giving the height of the aptamer above
the molecular backbone. Defaults to 3 mm. The aspect ratio of the aptamer is
fixed, so the width of the geom will be equal to the height.}
}
\description{
\code{geom_aptamer()} draws a geom representing an aptamer, a short
polynucleotide or polypeptide sequence that binds a target molecule.
}
\details{
The geom is drawn as a stylised secondary structure of a short
polynucleotide. The position of the aptamer on the molecular backbone is
expressed with the \code{x} aesthetic. Aptamers are not oriented, so this geom
does not accept the \code{forward} aesthetic. The aptamer is drawn with an
internal fill, which is white by default.
}
\section{Aesthetics}{

\itemize{
\item x (required; position of the aptamer)
\item y (required; the molecular backbone)
\item alpha
\item colour
\item linetype
\item size
\item fill
}
}

\examples{

ggplot2::ggplot(example_SBOL_features[example_SBOL_features$type == "aptamer", ],
                ggplot2::aes(x = start, y = molecule)) +
  geom_aptamer(inherit.aes = TRUE)

}
\seealso{
\code{\link[=geom_aptamer]{geom_aptamer()}}
}
