% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_aptamer_label.R
\name{geom_aptamer_label}
\alias{geom_aptamer_label}
\title{A 'ggplot' geom to add text labels to aptamers}
\usage{
geom_aptamer_label(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = FALSE,
  inherit.aes = FALSE,
  height = unit(4, "mm"),
  label_height = unit(3, "mm"),
  ...
)
}
\arguments{
\item{mapping, data, stat, position, na.rm, show.legend, inherit.aes, ...}{As
standard for ggplot2. inherit.aes is set to FALSE by default.}

\item{height}{\code{grid::unit()} object giving the height of the label above the
molecular backbone. Defaults to 4 mm.}

\item{label_height}{\code{grid::unit()} object giving the height of the label
text. Defaults to 3 mm.}
}
\description{
\code{geom_aptamer_label()} adds text labels to aptamers drawn with
\code{geom_aptamer()}.
}
\details{
Standard 'ggplot2' aesthetics for text are supported (see Aesthetics).
}
\section{Aesthetics}{

\itemize{
\item x (required; position of the aptamer)
\item y (required; the molecular backbone)
\item label (required; the label text)
\item forward
\item colour
\item size
\item alpha
\item family
\item fontface
\item angle
}
}

\examples{

ggplot2::ggplot(example_SBOL_features[example_SBOL_features$type == "aptamer", ],
                ggplot2::aes(x = start, y = molecule, label = name)) +
  geom_aptamer(inherit.aes = TRUE) +
  geom_aptamer_label(inherit.aes = TRUE)

}
\seealso{
\code{\link[=geom_aptamer]{geom_aptamer()}}
}
