% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_assembly_scar.R
\name{geom_assembly_scar}
\alias{geom_assembly_scar}
\title{A 'ggplot2' geom to draw assembly scars}
\usage{
geom_assembly_scar(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = FALSE,
  height = grid::unit(5, "mm"),
  ...
)
}
\arguments{
\item{mapping, data, stat, position, na.rm, show.legend, inherit.aes, ...}{As
standard for ggplot2. inherit.aes is set to FALSE by default.}

\item{height}{\code{grid::unit()} object giving the height that the assembly scar
extends away from the molecular backbone. Defaults to 5 mm.}
}
\description{
\code{geom_assembly_scar()} draws a geom representing an assembly scar, the
nucleotide sequence that is left behind after some cloning methods such as
restriction and ligation.
}
\details{
The assembly scar is drawn as two parallel lines offset outwards from the
molecular backbone. It has a fill between these lines which obscures the
backbone. This fill is white by default, but may need to be changed to match
the background colour of the plot.

The position of the assembly scar on the molecular backbone is set with the
\code{xmin} and \code{xmax} aesthetics. The molecular backbone that it is associated
with is set with the \code{y} aesthetic. This geom does not accept a forward
aesthetic, as it is not associated with a particular strand.
}
\section{Aesthetics}{

\itemize{
\item xmin,xmax (required; start and end positions of the assembly scar)
\item y (required; the molecular backbone)
\item forward
\item alpha
\item colour
\item linetype
\item size
\item fill
}
}

\examples{

ggplot2::ggplot(example_SBOL_features[example_SBOL_features$type == "assembly_scar", ],
                ggplot2::aes(xmin = start, xmax = end, y = molecule)) +
  geom_assembly_scar(inherit.aes = TRUE)

}
\seealso{
\code{\link[=geom_assembly_scar_label]{geom_assembly_scar_label()}}
}
