% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_assembly_scar_label.R
\name{geom_assembly_scar_label}
\alias{geom_assembly_scar_label}
\title{A 'ggplot' geom to add text labels to assembly scars}
\usage{
geom_assembly_scar_label(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = FALSE,
  inherit.aes = FALSE,
  height = unit(4, "mm"),
  label_height = unit(3, "mm"),
  ...
)
}
\arguments{
\item{mapping, data, stat, position, na.rm, show.legend, inherit.aes, ...}{As
standard for ggplot2. inherit.aes is set to FALSE by default.}

\item{height}{\code{grid::unit()} object giving the height of the label above the
molecular backbone. Defaults to 4 mm.}

\item{label_height}{\code{grid::unit()} object giving the height of the label
text. Defaults to 3 mm.}
}
\description{
\code{geom_assembly_scar_label()} adds text labels to assembly scars drawn with
\code{geom_assembly_scar()}.
}
\details{
Standard 'ggplot2' aesthetics for text are supported (see Aesthetics).
}
\section{Aesthetics}{

\itemize{
\item x (required; position of the assembly scar)
\item y (required; the molecular backbone)
\item label (required; the label text)
\item colour
\item size
\item alpha
\item family
\item fontface
\item angle
}
}

\examples{

ggplot2::ggplot(example_SBOL_features[example_SBOL_features$type == "assembly_scar", ],
                ggplot2::aes(xmin = start, xmax = end, y = molecule, label = name)) +
  geom_assembly_scar(inherit.aes = TRUE) +
  geom_assembly_scar_label(inherit.aes = TRUE)

}
\seealso{
\code{\link[=geom_assembly_scar]{geom_assembly_scar()}}
}
