% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_blunt_restriction_site.R
\name{geom_blunt_restriction_site}
\alias{geom_blunt_restriction_site}
\title{A 'ggplot2' geom to draw blunt restriction sites}
\usage{
geom_blunt_restriction_site(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = FALSE,
  height = grid::unit(5, "mm"),
  ...
)
}
\arguments{
\item{mapping, data, stat, position, na.rm, show.legend, inherit.aes, ...}{As
standard for ggplot2. inherit.aes is set to FALSE by default.}

\item{height}{\code{grid::unit()} opbject giving the height of the restriction
site above the molecular backbone. Defaults to 5 mm.}
}
\description{
\code{geom_blunt_restriction_site()} draws a geom representing a blunt
restriction site, the site at which a particular restriction enzyme cleaves
without leaving any overhanging nucleotides.
}
\details{
The blunt restriction site is drawn as a pair of straight brackets facing
away from one another, with the site location centered between the brackets.

The position of the restriction site on the molecular backbone is set with
the \code{x} aesthetic. The molecular backbone that it is associated with is set
with the \code{y} aesthetic. This geom does not accept a \code{forward} aesthetic, as
it is not associated with a particular strand.
}
\section{Aesthetics}{

\itemize{
\item x (required; position of the restriction site)
\item y (required; the molecular backbone)
\item alpha
\item colour
\item linetype
\item size
}
}

\examples{

ggplot2::ggplot(subset(feature_garden, 
                       feature == "blunt restriction site"),
  ggplot2::aes(x = start, y = molecule)) +
  geom_blunt_restriction_site(inherit.aes = TRUE)

}
\seealso{
\code{\link[=geom_blunt_restriction_site_label]{geom_blunt_restriction_site_label()}}
}
