% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_blunt_restriction_site_label.R
\name{geom_blunt_restriction_site_label}
\alias{geom_blunt_restriction_site_label}
\title{A 'ggplot' geom to add text labels to blunt restriction sites}
\usage{
geom_blunt_restriction_site_label(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = FALSE,
  inherit.aes = FALSE,
  height = unit(4, "mm"),
  label_height = unit(3, "mm"),
  ...
)
}
\arguments{
\item{mapping, data, stat, position, na.rm, show.legend, inherit.aes, ...}{As
standard for ggplot2. inherit.aes is set to FALSE by default.}

\item{height}{\code{grid::unit()} object giving the height of the label above the
molecular backbone. Defaults to 4 mm.}

\item{label_height}{\code{grid::unit()} object giving the height of the label
text. Defaults to 3 mm.}
}
\description{
\code{geom_blunt_restriction_site_label()} adds text labels to blunt restriction sites drawn with
\code{geom_blunt_restriction_site()}.
}
\details{
Standard 'ggplot2' aesthetics for text are supported (see Aesthetics).
}
\section{Aesthetics}{

\itemize{
\item x (required; position of the blunt restriction site)
\item y (required; the molecular backbone)
\item label (required; the label text)
\item colour
\item size
\item alpha
\item family
\item fontface
\item angle
}
}

\examples{

ggplot2::ggplot(feature_garden[feature_garden$feature == "blunt restriction site", ],
                ggplot2::aes(x = start, y = molecule, label = feature)) +
  geom_blunt_restriction_site(inherit.aes = TRUE) +
  geom_blunt_restriction_site_label(inherit.aes = TRUE)

}
\seealso{
\code{\link[=geom_blunt_restriction_site]{geom_blunt_restriction_site()}}
}
