% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_CDS.R
\name{geom_CDS}
\alias{geom_CDS}
\title{A 'ggplot2' geom to draw coding sequences (CDSs)}
\usage{
geom_CDS(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = FALSE,
  notched_arrow = FALSE,
  height = grid::unit(5, "mm"),
  ...
)
}
\arguments{
\item{mapping, data, stat, position, na.rm, show.legend, inherit.aes, ...}{As
standard for ggplot2. inherit.aes is set to FALSE by default.}

\item{notched_arrow}{If TRUE, the CDS is drawn as a 'notched' rather than
'box' arrow (FALSE by default)}

\item{height}{\code{grid::unit()} opbject giving the height of the CDS above
the molecular backbone. Defaults to 5 mm.}
}
\description{
\code{geom_CDS()} draws a geom representing a coding sequence, a nucleotide
sequence that encodes for a protein.
}
\details{
The CDS is drawn as an arrow, pointing in the 3′ direction of the strand it
is associated with. The arrow has a fill which is white by default.

The position of the CDS on the molecular backbone is set with the \code{xmin} and
\code{xmax} aesthetics. The molecular backbone that it is associated with is set
with the \code{y} aesthetic. The \code{forward} aesthetic can be used to set whether
the CDS is on the forward (default) or reverse strand. If \code{forward} is not
set, this will be inferred from the values of \code{xmin} and \code{xmax}.
}
\section{Aesthetics}{

\itemize{
\item xmin,xmax (required; start and end positions of the CDS)
\item y (required; the molecular backbone)
\item forward
\item alpha
\item colour
\item linetype
\item size
\item fill
}
}

\examples{

ggplot2::ggplot(subset(feature_garden, feature == "CDS"),
                ggplot2::aes(xmin = start, xmax = end, y = molecule)) +
  geom_CDS(inherit.aes = TRUE)

}
\seealso{
\code{\link[=geom_CDS_label]{geom_CDS_label()}}
}
