% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_CDS_label.R
\name{geom_CDS_label}
\alias{geom_CDS_label}
\title{A 'ggplot2' geom to add text labels to coding sequences (CDSs)}
\usage{
geom_CDS_label(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = FALSE,
  inherit.aes = FALSE,
  place = "inside",
  reverse_above = FALSE,
  height = grid::unit(4, "mm"),
  label_height = grid::unit(3, "mm"),
  align = "centre",
  ...
)
}
\arguments{
\item{mapping, data, stat, position, na.rm, show.legend, inherit.aes, ...}{As
standard for ggplot2. inherit.aes is set to FALSE by default.}

\item{place}{Where to draw the label, either 'inside' the CDS geom (the
default) or 'outside' (above the molecular backbone for CDSs on the forward
strand, and below it for CDSs on the reverse strand)}

\item{reverse_above}{If TRUE, and if place is 'outside', labels for CDSs on
the reverse strand will be drawn above the molecular backbone, as if they
are on the forward strand (FALSE by default)}

\item{height}{\code{grid::unit()} object giving the height of the label above the
molecular backbone, if place is 'outside'
forms. Defaults to 4 mm.}

\item{label_height}{\code{grid::unit()} object giving the height of the label
text. Defaults to 3 mm.}

\item{align}{How the text label should be aligned relative to the ends of
the CDS. Default is 'centre'; other options are 'left' and 'right'.}
}
\description{
\code{geom_CDS_label()} adds text labels to CDSs drawn with \code{geom_CDS()}.
}
\details{
Standard 'ggplot2' aesthetics for text are supported (see Aesthetics).
}
\section{Aesthetics}{

\itemize{
\item xmin,xmax (required; start and end positions of the CDS)
\item y (required; the molecular backbone)
\item label (required; the label text)
\item forward
\item colour
\item size
\item alpha
\item family
\item fontface
\item angle
}
}

\examples{

ggplot2::ggplot(subset(feature_garden, feature == "CDS"),
                ggplot2::aes(xmin = start, xmax = end, y = molecule, 
                             forward = forward, 
                             label = paste0(feature, ", ", variant))) +
  geom_CDS(inherit.aes = TRUE) +
  geom_CDS_label(inherit.aes = TRUE)

}
\seealso{
\code{\link[=geom_CDS]{geom_CDS()}}
}
