% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_cleavage_site.R
\name{geom_cleavage_site}
\alias{geom_cleavage_site}
\title{A 'ggplot2' geom to draw cleavage sites}
\usage{
geom_cleavage_site(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = FALSE,
  target = "DNA",
  reverse_above = FALSE,
  stem_height = grid::unit(3, "mm"),
  top_height = grid::unit(2, "mm"),
  ...
)
}
\arguments{
\item{mapping, data, stat, position, na.rm, show.legend, inherit.aes, ...}{As
standard for ggplot2. inherit.aes is set to FALSE by default.}

\item{target}{The target molecule, which determines how the stem is drawn.
One of 'DNA' (the default; straight line), 'RNA' (wavy line), or 'protein'
(looped line).}

\item{reverse_above}{If TRUE, cleavage sites on the reverse strand will be
drawn above the molecule backbone. FALSE by default.}

\item{stem_height}{\code{grid::unit()} object giving the height of the stem.
Defaults to 3 mm.}

\item{top_height}{\code{grid::unit()} object giving the height of the top.
Defaults to 2 mm. The aspect ratio of the top is fixed, so the height of the
top will also determine its width.}
}
\description{
\code{geom_cleavage_site()} draws a geom representing a cleavage site, for
example where a restriction enzyme, RNAse, or protease might act to cleave a
DNA, RNA, or protein molecule respectively.
}
\details{
The cleavage site is drawn as a 'stem-top' type geom, where the top
indicates the type of site (in this case, an 'X' representing a cleavage
site) and the step indicates the target molecule (a straight line for DNA,
wavy line for RNA, and looped line for protein).

The position of the cleavage site on the molecular backbone is set with the
\code{x} aesthetic. The molecular backbone that it is associated with is set with
the \code{y} aesthetic. The \code{forward} aesthetic can be used to set whether the
cleavage site is on the forward (default) or reverse strand.
}
\section{Aesthetics}{

\itemize{
\item x (required; position of the cleavage site)
\item y (required; the molecular backbone)
\item forward
\item alpha
\item colour
\item linetype
\item size
}
}

\examples{

ggplot2::ggplot(subset(feature_garden, feature == "cleavage site" &
                       variant == "DNA"),
                ggplot2::aes(x = start, y = molecule)) +
  geom_aptamer(inherit.aes = TRUE)

}
\seealso{
\code{\link[=geom_cleavage_site_label]{geom_cleavage_site_label()}}
}
