% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_cleavage_site_label.R
\name{geom_cleavage_site_label}
\alias{geom_cleavage_site_label}
\title{A 'ggplot2' geom to add text labels to cleavage sites}
\usage{
geom_cleavage_site_label(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = FALSE,
  inherit.aes = FALSE,
  reverse_above = FALSE,
  height = unit(4, "mm"),
  label_height = unit(3, "mm"),
  ...
)
}
\arguments{
\item{mapping, data, stat, position, na.rm, show.legend, inherit.aes, ...}{As
standard for ggplot2. inherit.aes is set to FALSE by default.}

\item{reverse_above}{If TRUE, labels on the reverse strand will be drawn
above the molecular backbone, as if they are on the forward strand. Defaults
to FALSE}

\item{height}{\code{grid::unit()} object giving the height of the label above the
molecular backbone. Defaults to 4 mm.}

\item{label_height}{\code{grid::unit()} object giving the height of the label
text. Defaults to 3 mm.}
}
\description{
\code{geom_cleavage_site_label()} adds text labels to cleavage sites drawn with
\code{geom_cleavage_site()}.
}
\details{
Standard 'ggplot2' aesthetics for text are supported (see Aesthetics).
}
\section{Aesthetics}{

\itemize{
\item x (required; position of the cleavage site)
\item y (required; the molecular backbone)
\item label (required; the label text)
\item forward
\item colour
\item size
\item alpha
\item family
\item fontface
\item angle
}
}

\examples{

ggplot2::ggplot(subset(feature_garden, feature == "cleavage site"),
                ggplot2::aes(x = start, y = molecule, label = feature)) +
  geom_cleavage_site(inherit.aes = TRUE) +
  geom_cleavage_site_label(inherit.aes = TRUE)

}
\seealso{
\code{\link[=geom_cleavage_site]{geom_cleavage_site()}}
}
