% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_engineered_region.R
\name{geom_engineered_region}
\alias{geom_engineered_region}
\title{A 'ggplot2' geom to draw engineered regions}
\usage{
geom_engineered_region(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = FALSE,
  reverse_above = FALSE,
  height = grid::unit(5, "mm"),
  ...
)
}
\arguments{
\item{mapping, data, stat, position, na.rm, show.legend, inherit.aes, ...}{As
standard for ggplot2. inherit.aes is set to FALSE by default.}

\item{reverse_above}{If TRUE, engineered regions on the reverse strand will
be drawn above the molecular backbone. FALSE by default}

\item{height}{\code{grid::unit()} object giving the height of the engineered region above
the molecular backbone. Defaults to 5 mm.}
}
\description{
\code{geom_engineered_region()} draws a geom representing an generic engineered
region of a sequence.
}
\details{
The engineered region is drawn as a plain rectangle suggesting a blank
slate. It has an internal fill, which is light grey by default.

The position of the engineered region on the molecular backbone is set with
the \code{xmin} and \code{xmax} aesthetics. The molecular backbone that it is
associated with is set with the \code{y} aesthetic. The \code{forward} aesthetic can
be used to set whether the engineered region is on the forward (default) or
reverse strand.
}
\section{Aesthetics}{

\itemize{
\item xmin/xmax (required; position of the engineered region)
\item y (required; the molecular backbone)
\item forward
\item alpha
\item colour
\item linetype
\item size
\item fill
}
}

\examples{

ggplot2::ggplot(subset(feature_garden, feature == "engineered region"),
                ggplot2::aes(xmin = start, xmax = end, forward = forward, 
                             y = molecule)) +
  geom_engineered_region(inherit.aes = TRUE)

}
\seealso{
\code{\link[=geom_engineered_region_label]{geom_engineered_region_label()}}
}
