% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_engineed_region_label.R
\name{geom_engineered_region_label}
\alias{geom_engineered_region_label}
\title{A 'ggplot2' geom to add text labels to engineered regions}
\usage{
geom_engineered_region_label(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = FALSE,
  inherit.aes = FALSE,
  place = "inside",
  reverse_above = FALSE,
  height = grid::unit(4, "mm"),
  label_height = grid::unit(5, "mm"),
  align = "centre",
  ...
)
}
\arguments{
\item{mapping, data, stat, position, na.rm, show.legend, inherit.aes, ...}{As
standard for ggplot2. inherit.aes is set to FALSE by default.}

\item{place}{Where to draw the label, either 'inside' the engineered region
geom (the default) or 'outside' (above the molecular backbone for engineered
regions on the forward strand, and below it for engineered regions on the
reverse strand)}

\item{reverse_above}{If TRUE, and if place is 'outside', labels for
engineered regions on the reverse strand will be drawn above the molecular
backbone, as if they are on the forward strand (FALSE by default)}

\item{height}{\code{grid::unit()} object giving the height of the label above the
molecular backbone, if place is 'outside'. Defaults to 4 mm.}

\item{label_height}{\code{grid::unit()} object giving the height of the label
text. Defaults to 5 mm.}

\item{align}{How the text label should be aligned relative to the ends of
the engineered region. Default is 'centre'; other options are 'left' and
'right'.}
}
\description{
\code{geom_engineered_region_label()} adds text labels to engineered regions
drawn with \code{geom_engineered_region()}.
}
\details{
Standard 'ggplot2' aesthetics for text are supported (see Aesthetics).
}
\section{Aesthetics}{

\itemize{
\item xmin,xmax (required; start and end positions of the engineered region)
\item y (required; the molecular backbone)
\item label (required; the label text)
\item forward
\item colour
\item size
\item alpha
\item family
\item fontface
\item angle
}
}

\examples{

ggplot2::ggplot(subset(feature_garden, feature == "engineered region"),
                ggplot2::aes(xmin = start, xmax = end, y = molecule, 
                             forward = forward, 
                             label = paste0(feature, ", ", variant))) +
  geom_engineered_region(inherit.aes = TRUE) +
  geom_engineered_region_label(inherit.aes = TRUE)

}
\seealso{
\code{\link[=geom_engineered_region]{geom_engineered_region()}}
}
