% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_subgene_arrow.R
\name{geom_subgene_arrow}
\alias{geom_subgene_arrow}
\title{`geom_subgene_arrow` draws subgenes segments as arrow fragments, allowing segments of arrows to overlay gene maps drawings}
\usage{
geom_subgene_arrow(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, arrowhead_width = grid::unit(4, "mm"),
  arrowhead_height = grid::unit(4, "mm"),
  arrow_body_height = grid::unit(3, "mm"), ...)
}
\arguments{
\item{mapping, data, stat, position, na.rm, show.legend, inherit.aes, ...}{As
standard for ggplot2.}

\item{arrowhead_width}{grid::unit object giving the width of the arrowhead.
Defaults to 4 mm. If the gene is drawn smaller than this width, only the
arrowhead will be drawn, compressed to the length of the gene.}

\item{arrowhead_height}{grid::unit object giving the height of the arrowhead.
Defaults to 4 mm.}

\item{arrow_body_height}{grid::unit object giving the height of the body of
the arrow. Defaults to 3 mm.}
}
\description{
This geom draws gene segments as arrows along a horizontal line representing
the molecule. The start and end locations of the gene are expressed with the
`xmin` and `xmax` aesthetics, while the molecule can be specified with the
`y` aesthetic. The `xsubmin` and `xsubmax` aesthetics must also be supplied,
representing the subgene segments that should be drawn. The subgene segments
will be drawn w.r.t. the arrowhead of the underlying genes, so `xmin` and
`xmax` must also be given, as should other aesthetics and arguments that
could be used in \code{\link{geom_gene_arrow}} (`forward`,
`arrowhead_width`, etc).
}
\section{Aesthetics}{


\itemize{
  \item xmin,xmax (start and end of the gene; will be used to determine
        gene orientation)
  \item xsubmin,xsubmax (start and end of subgene segment). Should be consistent with `xmin`/`xmax`.
  \item y (molecule)
  \item forward (if any value that is not TRUE, or coercible to TRUE, the
  gene arrow will be drawn in the opposite direction to that determined by
  `xmin` and `xmax`)
  \item alpha
  \item colour
  \item fill
  \item linetype
  \item size
}
}

\examples{

ggplot2::ggplot(example_genes, ggplot2::aes(xmin = start, xmax = end,
                                            y = molecule)) +
geom_gene_arrow() +
geom_subgene_arrow(data = example_subgenes,
      ggplot2::aes(xmin = start, xmax = end, xsubmin = from, xsubmax = to,
                   y = molecule, fill = gene)) +
ggplot2::facet_wrap(~ molecule, scales = "free")

}
\seealso{
theme_genes, make_alignment_dummies, geom_gene_arrow, geom_gene_label
}
