% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gggenes-package.R
\docType{package}
\name{gggenes-package}
\alias{gggenes}
\alias{gggenes-package}
\title{gggenes: Draw Gene Arrow Maps in 'ggplot2'}
\description{
A 'ggplot2' extension for drawing gene arrow maps.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://wilkox.org/gggenes/}
  \item Report bugs at \url{https://github.com/wilkox/gggenes/issues/}
}

}
\author{
\strong{Maintainer}: David Wilkins \email{david@wilkox.org}

Other contributors:
\itemize{
  \item Zachary Kurtz \email{zdkurtz@gmail.com} [contributor]
}

}
\keyword{internal}
