% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gggenes.R
\docType{package}
\name{gggenes}
\alias{gggenes}
\alias{gggenes-package}
\title{'gggenes': provides a 'ggplot2' geom and helper functions for drawing gene
arrow maps.}
\description{
This package provides a 'ggplot2' geom, `geom_gene_arrow` a theme,
`theme_genes`, and a helper function for visually aligning genes,
`make_alignment_dummies`.
}
