% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_alignment_dummies.R
\name{prepare_alignment_dummies}
\alias{prepare_alignment_dummies}
\title{Prepare dummies to align genes across faceted molecules}
\usage{
prepare_alignment_dummies(data, mapping, align.on, align.side = "left")
}
\arguments{
\item{data}{Data frame. Probably the same data frame passed to
\code{ggplot2}, \code{geom_gene_arrow} etc.}

\item{mapping}{Mapping created with \code{aes}. Must contain the following
aesthetics: xmin, xmax, y, and id. ‘id’ is the column containing gene ids, of
which the ‘align.on’ gene must be one.}

\item{align.on}{Name of gene to align on. Must be present in ‘data’, in the
column mapped to ‘id’.}

\item{align.side}{Align to either the ‘left’ (default) or ‘right’ side of the
target gene.}
}
\description{
Given a gene to align on, \code{prepare_alignment_dummies} will produce a
data frame of dummies that can be added to a gene map with \code{geom_blank}.
These dummies will visually align the start (or end) of the target gene
across molecules, and will also ensure that the x-axis range is uniform
across all molecules.
}
\examples{

dummies <- prepare_alignment_dummies(gene_data, aes(xmin = start, xmax = end,
y = contig, id = gene_id), align.on = "nifB")

ggplot(gene_data, aes(xmin = start, xmax = end, y = contig)) +
  geom_gene_arrow() +
  geom_blank(data = dummies, aes(xmin = dummy_start, xmax = dummy_end, y =
  contig), inherit.aes = F) +
  facet_wrap(~ Contig, scales = "free", ncol = 1) +
  theme_genes()
}

