% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_genes.R
\name{theme_genes}
\alias{theme_genes}
\title{Theme for gene maps.}
\usage{
theme_genes(...)
}
\description{
A ggplot2 theme optimised for drawing gene maps.

This theme removes strip text (the text that labels facets when you use
\code{facet_wrap} or \code{facet_grid}. This is to allow for easy
drawing of molecules on different x-scales by setting the y aesthetic to the
molecule, then faceting with `facet_grid( ~ molecule, scales = "free")`.
}

