% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_genes}
\alias{example_genes}
\alias{example_subgenes}
\title{A set of example genes.}
\format{A data frame with 118 rows and four variables:
\describe{
\item{molecule}{the genome}
\item{gene}{the name of the gene}
\item{start}{the start position of the gene}
\item{end}{the end position of the gene}
\item{strand}{the strand of the gene}
}

example_subgenes (237 rows) also contains:
\describe{
\item{subgeme}{the name of the subgene}
\item{from}{the start position of the subgene segment}
\item{strand}{the end position of the subgene segment}
}}
\usage{
example_genes

example_subgenes
}
\description{
Genes for example purposes only.
}
\keyword{datasets}
