% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_genes}
\alias{example_genes}
\alias{example_genes_polar}
\alias{example_subgenes}
\alias{example_subgenes_polar}
\title{A set of example genes.}
\format{
A data frame with 72 rows and six variables:
\describe{
\item{molecule}{the genome}
\item{gene}{the name of the gene}
\item{start}{the start position of the gene}
\item{end}{the end position of the gene}
\item{strand}{the strand of the gene}
\item{orientation}{the orientation of the gene}
}

example_subgenes (143 rows) also contains:
\describe{
\item{subgene}{the name of the subgene}
\item{from}{the start position of the subgene segment}
\item{to}{the end position of the subgene segment}
}

An object of class \code{data.frame} with 18 rows and 6 columns.

An object of class \code{data.frame} with 143 rows and 9 columns.

An object of class \code{data.frame} with 40 rows and 9 columns.
}
\usage{
example_genes

example_genes_polar

example_subgenes

example_subgenes_polar
}
\description{
Genes for example purposes only.
}
\keyword{datasets}
