% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_SBOL_features}
\alias{example_SBOL_features}
\title{A set of example SBOL sequence features}
\format{
A data frame with six rows and three variables:
\describe{
\item{molecule}{the genome}
\item{start}{the start position or location of the feature}
\item{type}{the type of sequence features}
\item{name}{the name of the feature}
}
}
\usage{
example_SBOL_features
}
\description{
Sequence features for example purposes only.
}
\keyword{datasets}
