% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_CDS_label.R
\name{geom_CDS_label}
\alias{geom_CDS_label}
\title{A 'ggplot2' geom to add text labels to coding sequences (CDSs)}
\usage{
geom_CDS_label(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = FALSE,
  inherit.aes = FALSE,
  height = grid::unit(4, "mm"),
  label_height = grid::unit(3, "mm"),
  variant = "default",
  align = "centre",
  ...
)
}
\arguments{
\item{mapping, data, stat, position, na.rm, show.legend, inherit.aes, ...}{As
standard for ggplot2. inherit.aes is set to FALSE by default.}

\item{height}{\code{grid::unit()} object giving the height of the label above the
molecular backbone, if drawn with the 'above' or 'reverse_above' variant
forms. Defaults to 4 mm.}

\item{label_height}{\code{grid::unit()} object giving the height of the label
text. Defaults to 3 mm.}

\item{variant}{Specify a variant form of the geom (see section Variant
forms).}

\item{align}{Where inside the gene to place the text label. Default is
'centre'; other options are 'left' and 'right'.}
}
\description{
\code{geom_CDS_label()} adds text labels to CDSs drawn with \code{geom_CDS()}.
}
\details{
Standard 'ggplot2' aesthetics for text are supported (see Aesthetics).
}
\section{Variant forms}{

\itemize{
\item default: the default form, where the label is drawn inside the CDS geom
\item above: labels are drawn outside of the CDS geom, above the molecular
backbone for CDSs on the forward strand and below it for CDSs on the reverse
strand
\item reverse_above: labels for CDSs on the reverse strand will be drawn above
the molecular backbone, as if they are on the forward strand
}
}

\section{Aesthetics}{

\itemize{
\item xmin,xmax (required; start and end positions of the CDS)
\item y (required; the molecular backbone)
\item label (required; the label text)
\item forward
\item colour
\item size
\item alpha
\item family
\item fontface
\item angle
}
}

\examples{

ggplot2::ggplot(subset(feature_garden, feature == "CDS"),
                ggplot2::aes(xmin = start, xmax = end, y = molecule, 
                             forward = forward, 
                             label = paste0(feature, ", ", variant))) +
  geom_CDS(inherit.aes = TRUE) +
  geom_CDS_label(inherit.aes = TRUE)

}
\seealso{
\code{\link[=geom_CDS]{geom_CDS()}}
}
