% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_circular_plasmid.R
\name{geom_circular_plasmid}
\alias{geom_circular_plasmid}
\title{A 'ggplot2' geom to indicate a circular plasmid section}
\usage{
geom_circular_plasmid(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = FALSE,
  place = "default",
  ...
)
}
\arguments{
\item{mapping, data, stat, position, na.rm, show.legend, inherit.aes, ...}{As
standard for ggplot2. inherit.aes is set to FALSE by default.}

\item{place}{If 'default' (the default), a pair of 'left' and 'right' curves
(if \code{xmin}/\code{xmax} are mapped) or a single 'left' curve (if \code{x} is mapped).
If 'left', a single 'left' curve, with the location set by either \code{xmin} or
\code{x}. If 'right', a single 'right' curve, with the location set by either
\code{xmax} or \code{x}.}
}
\description{
\code{geom_circular_plasmid()} draws a geom indicating that the section of
molecule being depicted is part of a circular plasmid.
}
\details{
The geom is a hook-shaped curve or pair of curves coming off the molecular
backbone.

By default, the geom comprises a pair of 'left' and 'right' curves, with the
\code{xmin} and \code{xmax} aesthetics used to set their positions. A single 'left' or
'right' curve can be drawn by selecting the 'left' or 'right' place, and/or
setting the \code{x} aesthetic instead of \code{xmin}/\code{xmax}.
}
\section{Aesthetics}{

\itemize{
\item x or xmin,xmax (required; one or both ends of the plasmid section)
\item y (required; the molecular backbone)
\item alpha
\item color
\item linetype
\item size
\item fill
}
}

\examples{

ggplot2::ggplot(subset(feature_garden, feature == "circular plasmid" &
                        variant == "default"),
                ggplot2::aes(xmin = start, xmax = end, y = molecule)) +
  geom_circular_plasmid(inherit.aes = TRUE)

}
\seealso{
\code{\link[=geom_circular_plasmid]{geom_circular_plasmid()}}
}
