% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_circular_plasmid_label.R
\name{geom_circular_plasmid_label}
\alias{geom_circular_plasmid_label}
\title{A 'ggplot2' geom to add a text label to a circular plasmid section}
\usage{
geom_circular_plasmid_label(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = FALSE,
  inherit.aes = FALSE,
  height = unit(4, "mm"),
  label_height = unit(3, "mm"),
  place = "left",
  ...
)
}
\arguments{
\item{mapping, data, stat, position, na.rm, show.legend, inherit.aes, ...}{As
standard for ggplot2. inherit.aes is set to FALSE by default.}

\item{height}{\code{grid::unit()} object giving the height of the label above the
molecular backbone. Defaults to 4 mm.}

\item{label_height}{\code{grid::unit()} object giving the height of the label
text. Defaults to 3 mm.}

\item{place}{Where to place the label, one of left (default; over the left
end of the locus, or the right end if there is no left), right (over the
right end of the locus, or the left end if there is no right), or middle
(centred over the locus if it has two ends, or over the end if it has one
end)}
}
\description{
\code{geom_circular_plasmid_label()} adds text labels to circular plasmid
sections drawn with \code{geom_circular_plasmid()}.
}
\details{
Standard 'ggplot2' aesthetics for text are supported (see Aesthetics). For
compatibility, the geom accepts the same combinations of \code{x} or
\code{xmin}/\code{xmax} aesthetics.
}
\section{Aesthetics}{

\itemize{
\item x or xmin,xmax (required; one or both ends of the section)
\item y (required; the molecular backbone)
\item label (required; the label text)
\item colour
\item size
\item alpha
\item family
\item fontface
\item angle
}
}

\examples{

ggplot2::ggplot(subset(feature_garden, feature == "circular plasmid" &
                        variant == "default"),
                ggplot2::aes(xmin = start, xmax = end, y = molecule,
                             label = feature)) +
  geom_circular_plasmid(inherit.aes = TRUE) +
  geom_circular_plasmid_label(inherit.aes = TRUE)

}
\seealso{
\code{\link[=geom_circular_plasmid]{geom_circular_plasmid()}}
}
