% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_gene_arrow.R
\name{geom_gene_arrow}
\alias{geom_gene_arrow}
\title{Draw a gene as an arrow.}
\usage{
geom_gene_arrow(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, arrowhead_width = unit(4, "mm"),
  arrowhead_height = unit(4, "mm"), arrow_body_height = unit(3, "mm"), ...)
}
\arguments{
\item{mapping, data, stat, identity, na.rm, show.legend, inherit.aes, ...}{As
standard for ggplot2.}

\item{arrowhead_width}{Unit object giving the width of the arrowhead.
Defaults to 4 mm. If the gene is drawn smaller than this width, only the
arrowhead will be drawn, compressed to the length of the gene.}

\item{arrowhead_height}{Unit object giving the height of the arrowhead.
Defaults to 4 mm.}

\item{arrow_body_height}{Unit object giving the height of the body of the
arrow. Defaults to 3 mm.}
}
\description{
Draws an arrow representing a gene with orientation. Only works for genes
arranged horizontally, i.e. with the x-axis as molecule position and the
y-axis as molecule.

\code{geom_gene_arrow} understands the following aesthetics:
\itemize{
\item{xmin and xmax (start and end of the gene; will be used to determine
gene orientation)}
\item{y}
\item{alpha}
\item{colour}
\item{fill}
\item{linetype}
}
}

