% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_genes.R
\docType{data}
\name{theme_genes}
\alias{theme_genes}
\title{Theme for gene maps.}
\format{An object of class \code{theme} (inherits from \code{gg}) of length 63.}
\usage{
theme_genes
}
\description{
A ggplot2 theme optimised for drawing gene maps.
}
\keyword{datasets}

